+++ releasenotes.md
@@ -1,2 +1,3 @@
 Whats New:
 ----------------------
+- We added a functionality to create ingredients replacements data